% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cea.r
\name{cea}
\alias{cea}
\title{Simulate model with multiple input.}
\usage{
cea(mfunc, md, pars.tab0, pars.tab1, yini, rho_cost, rho_qol, fr, to, dt = 1,
  n.sim = 100)
}
\arguments{
\item{md}{model definition}

\item{pars.tab0}{list of probabilistic distributions of parameters for standard treatment}

\item{pars.tab1}{list of probabilistic distributions of parameters for new intervention}

\item{yini}{initial value for simulation}

\item{rho_cost}{interest rate for future cost}

\item{rho_qol}{interst rate for future qol}

\item{fr}{start time}

\item{to}{end time}

\item{dt}{observation interval}

\item{n.sim}{size of parameter sets}

\item{nfunc}{Constructor of model, (DetMC, StoMC, ODE, IBM)}
}
\value{
object of cea
}
\description{
Simulate model with multiple input.
}
\examples{


}

