% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm.r
\name{check.parameters}
\alias{check.parameters}
\title{Check the values of parameters
print the ill-defined parameters and parameters applied default values}
\usage{
check.parameters(dm, pars)
}
\arguments{
\item{dm}{definition sheet of model}

\item{pars}{a vector or list of the model}
}
\value{
True if checking successful
}
\description{
Check the values of parameters
print the ill-defined parameters and parameters applied default values
}
\examples{

}

