% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm.r
\name{ddm}
\alias{ddm}
\title{Extract information in model definition script.}
\usage{
ddm(ss)
}
\arguments{
\item{ss}{script}
}
\value{
list of \itemize{
  \item \code{dot} {string in DOT language for export and visualization}
  \item \code{dotc} {string in DOT language for export and visualization (cost/qol node info dropped for conceptual overview)}
  \item \code{model.def} {Definition of dynamic model}
  \item \code{(todo) parmeters}{list of parameters}
}
}
\description{
The algorithm was a modified version of dtree package (https://github.com/petedodd/dtree/)
}
\examples{
 script0 <- "
 \\\\nodes...(this line is  comment)
   \\\\ nodename ['label | cost variable | QoL variable']
   A[ 'description A | costA | qolA'] \\\\ a node
   B[ 'description B | costB | qolB']
 C[  'description C | costC | qolC']
 \\\\edges
 \\\\ from_statename -> to_statename['transition function or name for this edge']
 A -> B ['TrAB']
 A -> C ['TrAC']
 B -> A ['TrBC']

 \\\\ (whitespace is ignored in parsing for calculation)
 "

 md = ddm(script0)

}

