% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detmc.r
\name{DetMC}
\alias{DetMC}
\title{Deterministic Markov Chain Model
create a deterministic Markov model. Basicly, the model is defined under continuous time}
\usage{
DetMC(dm, pars, default.cost = 1, default.qol = 1, default.trs = 1)
}
\arguments{
\item{dm}{definiation of model}

\item{pars}{list or vector of parameters}

\item{default.cost}{default value of cost per simulation interval per person}

\item{default.qol}{default value of qol per simulation interval per person, ranged [0, 1]}

\item{default.trs}{default value of intensity (transition rate)}
}
\value{
Deterministic Markov Chain Model for simulation
}
\description{
Deterministic Markov Chain Model
create a deterministic Markov model. Basicly, the model is defined under continuous time
}
\examples{

}

