% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm.r
\name{goto}
\alias{goto}
\title{Simulate forward one unit}
\usage{
goto(model, y, ti, dt)
}
\arguments{
\item{model}{dynamic model}

\item{y}{initial value}

\item{ti}{initial time}

\item{dt}{step size}
}
\value{
y after simulation
}
\description{
Simulate forward one unit
}
\examples{

}

