% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibm.r
\name{IBM}
\alias{IBM}
\title{Individual-Based Model (microsimulation)
create an individual based model. The state}
\usage{
IBM(md, pars, default.cost = 1, default.qol = 1, default.trs = 1,
  ini.attr = F)
}
\arguments{
\item{pars}{list or vector of parameters}

\item{default.cost}{default value of cost per simulation interval per person}

\item{default.qol}{default value of qol per simulation interval per person, ranged [0, 1]}

\item{default.trs}{default value of intensity (transition rate)}

\item{ini.attr}{function for generating attributes of an agent ()=>vector(Attr: Val)}

\item{dm}{definiation of model}
}
\value{
individual based model for simulation
}
\description{
Individual-Based Model (microsimulation)
create an individual based model. The state
}
\examples{

}

