% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty.r
\name{make.parameters}
\alias{make.parameters}
\title{Parameter table maker}
\usage{
make.parameters(pars.sim, n.sim)
}
\arguments{
\item{pars.sim}{list of probabilistic distributions of parameters, double value if constant}

\item{n.sim}{size of parameter sets}
}
\value{
data.frame of parameter table
}
\description{
Parameter table maker
}
\examples{
pars.def <- list(TrAB=0.02, TrBC='exp(0.2)', TrAC='gamma(0.01, 0.1)')
make.parameters(pars.def, 10)

}

