% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode.r
\name{ODE}
\alias{ODE}
\title{Ordinary differential equation Model
create an ODE model. Basicly, the model is defined under continuous time}
\usage{
ODE(m, pars, default.cost = 1, default.qol = 1, default.trs = 1)
}
\arguments{
\item{pars}{list or vector of parameters}

\item{default.cost}{default value of cost per simulation interval per person}

\item{default.qol}{default value of qol per simulation interval per person, ranged [0, 1]}

\item{default.trs}{default value of intensity (transition rate)}

\item{dm}{definiation of model}
}
\value{
ODE for simulation
}
\description{
Ordinary differential equation Model
create an ODE model. Basicly, the model is defined under continuous time
}
\examples{

}

