% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm.r
\name{simHE}
\alias{simHE}
\title{Simulate observations from dynamic model}
\usage{
simHE(model, yini, from, to, dt = 1, rec = "m", ...)
}
\arguments{
\item{model}{dynamic model}

\item{yini}{initial values}

\item{from}{double, start time}

\item{to}{double, end time}

\item{dt}{double, observation interval}

\item{rec}{characher, timing for observation, (s: starting, m: middle, e: ending)}
}
\value{
\code{Y}: last value in the end of simulation

\code{Obs}: the observed time series (data.frame)
}
\description{
Simulate observations from dynamic model
}
\examples{

}

