% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stomc.r
\name{StoMC}
\alias{StoMC}
\title{Stochastic Markov Chain Model
create a Stochastic Markov model. Basicly, the model is defined under continuous time}
\usage{
StoMC(m, pars, default.cost = 1, default.qol = 1, default.trs = 1)
}
\arguments{
\item{pars}{list or vector of parameters}

\item{default.cost}{default value of cost per simulation interval per person}

\item{default.qol}{default value of qol per simulation interval per person, ranged [0, 1]}

\item{default.trs}{default value of intensity (transition rate)}

\item{dm}{definiation of model}
}
\value{
Stochastic Markov Chain Model for simulation
}
\description{
Stochastic Markov Chain Model
create a Stochastic Markov model. Basicly, the model is defined under continuous time
}
\examples{

}

