% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cea.r
\name{summarise.CE}
\alias{summarise.CE}
\title{Summarise cost and qol with time preference}
\usage{
summarise.CE(out, rho_cost = 0.02, rho_qol = 0.02)
}
\arguments{
\item{out}{output table from simulation model}

\item{rho_cost}{interest rate for future cost}

\item{rho_qol}{interst rate for future qol}
}
\value{
a vector of cost and qol
}
\description{
Summarise cost and qol with time preference
}
\examples{

}

