% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm.r
\name{observe}
\alias{observe}
\title{Make observations of dynamic model}
\usage{
observe(model, y, ti, ...)
}
\arguments{
\item{model}{dynamic model}

\item{y}{status of the model}

\item{ti}{current time}
}
\value{
vector of obasevations
}
\description{
Make observations of dynamic model
}
\examples{

}

